+++ cams-20200606/alibabacloud_cams20200606/models/_chatapp_sync_phone_number_response_body.py
@@ -97,6 +97,7 @@ class ChatappSyncPhoneNumberResponseBody(DaraModel):
 class ChatappSyncPhoneNumberResponseBodyPhoneNumbers(DaraModel):
     def __init__(
         self,
+        calling_configure: main_models.ChatappSyncPhoneNumberResponseBodyPhoneNumbersCallingConfigure = None,
         code_verification_status: str = None,
         is_official: str = None,
         messaging_limit_tier: str = None,
@@ -111,6 +112,7 @@ class ChatappSyncPhoneNumberResponseBodyPhoneNumbers(DaraModel):
         up_queue: str = None,
         verified_name: str = None,
     ):
+        self.calling_configure = calling_configure
         # The verification state of the phone number.
         # 
         # Valid values:
@@ -166,13 +168,17 @@ class ChatappSyncPhoneNumberResponseBodyPhoneNumbers(DaraModel):
         self.verified_name = verified_name
 
     def validate(self):
-        pass
+        if self.calling_configure:
+            self.calling_configure.validate()
 
     def to_map(self):
         result = dict()
         _map = super().to_map()
         if _map is not None:
             result = _map
+        if self.calling_configure is not None:
+            result['CallingConfigure'] = self.calling_configure.to_map()
+
         if self.code_verification_status is not None:
             result['CodeVerificationStatus'] = self.code_verification_status
 
@@ -216,6 +222,10 @@ class ChatappSyncPhoneNumberResponseBodyPhoneNumbers(DaraModel):
 
     def from_map(self, m: dict = None):
         m = m or dict()
+        if m.get('CallingConfigure') is not None:
+            temp_model = main_models.ChatappSyncPhoneNumberResponseBodyPhoneNumbersCallingConfigure()
+            self.calling_configure = temp_model.from_map(m.get('CallingConfigure'))
+
         if m.get('CodeVerificationStatus') is not None:
             self.code_verification_status = m.get('CodeVerificationStatus')
 
@@ -254,6 +264,258 @@ class ChatappSyncPhoneNumberResponseBodyPhoneNumbers(DaraModel):
 
         if m.get('VerifiedName') is not None:
             self.verified_name = m.get('VerifiedName')
+
+        return self
+
+class ChatappSyncPhoneNumberResponseBodyPhoneNumbersCallingConfigure(DaraModel):
+    def __init__(
+        self,
+        calling: main_models.ChatappSyncPhoneNumberResponseBodyPhoneNumbersCallingConfigureCalling = None,
+        calling_callback_url: str = None,
+        max_talk_time: int = None,
+    ):
+        self.calling = calling
+        self.calling_callback_url = calling_callback_url
+        self.max_talk_time = max_talk_time
+
+    def validate(self):
+        if self.calling:
+            self.calling.validate()
+
+    def to_map(self):
+        result = dict()
+        _map = super().to_map()
+        if _map is not None:
+            result = _map
+        if self.calling is not None:
+            result['Calling'] = self.calling.to_map()
+
+        if self.calling_callback_url is not None:
+            result['CallingCallbackUrl'] = self.calling_callback_url
+
+        if self.max_talk_time is not None:
+            result['MaxTalkTime'] = self.max_talk_time
+
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('Calling') is not None:
+            temp_model = main_models.ChatappSyncPhoneNumberResponseBodyPhoneNumbersCallingConfigureCalling()
+            self.calling = temp_model.from_map(m.get('Calling'))
+
+        if m.get('CallingCallbackUrl') is not None:
+            self.calling_callback_url = m.get('CallingCallbackUrl')
+
+        if m.get('MaxTalkTime') is not None:
+            self.max_talk_time = m.get('MaxTalkTime')
+
+        return self
+
+class ChatappSyncPhoneNumberResponseBodyPhoneNumbersCallingConfigureCalling(DaraModel):
+    def __init__(
+        self,
+        call_hours: main_models.ChatappSyncPhoneNumberResponseBodyPhoneNumbersCallingConfigureCallingCallHours = None,
+        call_icon_visibility: str = None,
+        callback_permission_status: str = None,
+        status: str = None,
+    ):
+        self.call_hours = call_hours
+        self.call_icon_visibility = call_icon_visibility
+        self.callback_permission_status = callback_permission_status
+        self.status = status
+
+    def validate(self):
+        if self.call_hours:
+            self.call_hours.validate()
+
+    def to_map(self):
+        result = dict()
+        _map = super().to_map()
+        if _map is not None:
+            result = _map
+        if self.call_hours is not None:
+            result['CallHours'] = self.call_hours.to_map()
+
+        if self.call_icon_visibility is not None:
+            result['CallIconVisibility'] = self.call_icon_visibility
+
+        if self.callback_permission_status is not None:
+            result['CallbackPermissionStatus'] = self.callback_permission_status
+
+        if self.status is not None:
+            result['Status'] = self.status
+
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('CallHours') is not None:
+            temp_model = main_models.ChatappSyncPhoneNumberResponseBodyPhoneNumbersCallingConfigureCallingCallHours()
+            self.call_hours = temp_model.from_map(m.get('CallHours'))
+
+        if m.get('CallIconVisibility') is not None:
+            self.call_icon_visibility = m.get('CallIconVisibility')
+
+        if m.get('CallbackPermissionStatus') is not None:
+            self.callback_permission_status = m.get('CallbackPermissionStatus')
+
+        if m.get('Status') is not None:
+            self.status = m.get('Status')
+
+        return self
+
+class ChatappSyncPhoneNumberResponseBodyPhoneNumbersCallingConfigureCallingCallHours(DaraModel):
+    def __init__(
+        self,
+        holiday_schedule: List[main_models.ChatappSyncPhoneNumberResponseBodyPhoneNumbersCallingConfigureCallingCallHoursHolidaySchedule] = None,
+        status: str = None,
+        timezone_id: str = None,
+        weekly_operating_hours: List[main_models.ChatappSyncPhoneNumberResponseBodyPhoneNumbersCallingConfigureCallingCallHoursWeeklyOperatingHours] = None,
+    ):
+        self.holiday_schedule = holiday_schedule
+        self.status = status
+        self.timezone_id = timezone_id
+        self.weekly_operating_hours = weekly_operating_hours
+
+    def validate(self):
+        if self.holiday_schedule:
+            for v1 in self.holiday_schedule:
+                 if v1:
+                    v1.validate()
+        if self.weekly_operating_hours:
+            for v1 in self.weekly_operating_hours:
+                 if v1:
+                    v1.validate()
+
+    def to_map(self):
+        result = dict()
+        _map = super().to_map()
+        if _map is not None:
+            result = _map
+        result['HolidaySchedule'] = []
+        if self.holiday_schedule is not None:
+            for k1 in self.holiday_schedule:
+                result['HolidaySchedule'].append(k1.to_map() if k1 else None)
+
+        if self.status is not None:
+            result['Status'] = self.status
+
+        if self.timezone_id is not None:
+            result['TimezoneId'] = self.timezone_id
+
+        result['WeeklyOperatingHours'] = []
+        if self.weekly_operating_hours is not None:
+            for k1 in self.weekly_operating_hours:
+                result['WeeklyOperatingHours'].append(k1.to_map() if k1 else None)
+
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        self.holiday_schedule = []
+        if m.get('HolidaySchedule') is not None:
+            for k1 in m.get('HolidaySchedule'):
+                temp_model = main_models.ChatappSyncPhoneNumberResponseBodyPhoneNumbersCallingConfigureCallingCallHoursHolidaySchedule()
+                self.holiday_schedule.append(temp_model.from_map(k1))
+
+        if m.get('Status') is not None:
+            self.status = m.get('Status')
+
+        if m.get('TimezoneId') is not None:
+            self.timezone_id = m.get('TimezoneId')
+
+        self.weekly_operating_hours = []
+        if m.get('WeeklyOperatingHours') is not None:
+            for k1 in m.get('WeeklyOperatingHours'):
+                temp_model = main_models.ChatappSyncPhoneNumberResponseBodyPhoneNumbersCallingConfigureCallingCallHoursWeeklyOperatingHours()
+                self.weekly_operating_hours.append(temp_model.from_map(k1))
+
+        return self
+
+class ChatappSyncPhoneNumberResponseBodyPhoneNumbersCallingConfigureCallingCallHoursWeeklyOperatingHours(DaraModel):
+    def __init__(
+        self,
+        close_time: str = None,
+        day_of_week: str = None,
+        open_time: str = None,
+    ):
+        self.close_time = close_time
+        self.day_of_week = day_of_week
+        self.open_time = open_time
+
+    def validate(self):
+        pass
+
+    def to_map(self):
+        result = dict()
+        _map = super().to_map()
+        if _map is not None:
+            result = _map
+        if self.close_time is not None:
+            result['CloseTime'] = self.close_time
+
+        if self.day_of_week is not None:
+            result['DayOfWeek'] = self.day_of_week
+
+        if self.open_time is not None:
+            result['OpenTime'] = self.open_time
+
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('CloseTime') is not None:
+            self.close_time = m.get('CloseTime')
+
+        if m.get('DayOfWeek') is not None:
+            self.day_of_week = m.get('DayOfWeek')
+
+        if m.get('OpenTime') is not None:
+            self.open_time = m.get('OpenTime')
+
+        return self
+
+class ChatappSyncPhoneNumberResponseBodyPhoneNumbersCallingConfigureCallingCallHoursHolidaySchedule(DaraModel):
+    def __init__(
+        self,
+        date: str = None,
+        end_time: str = None,
+        start_time: str = None,
+    ):
+        self.date = date
+        self.end_time = end_time
+        self.start_time = start_time
+
+    def validate(self):
+        pass
+
+    def to_map(self):
+        result = dict()
+        _map = super().to_map()
+        if _map is not None:
+            result = _map
+        if self.date is not None:
+            result['Date'] = self.date
+
+        if self.end_time is not None:
+            result['EndTime'] = self.end_time
+
+        if self.start_time is not None:
+            result['StartTime'] = self.start_time
+
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('Date') is not None:
+            self.date = m.get('Date')
+
+        if m.get('EndTime') is not None:
+            self.end_time = m.get('EndTime')
+
+        if m.get('StartTime') is not None:
+            self.start_time = m.get('StartTime')
 
         return self