+++ cams-20200606/ChangeLog.md
@@ -1,3 +1,10 @@
+2026-02-06 Version: 4.0.2
+- Update API CreateFlow: add request parameters EndpointUri.
+- Update API GetFlow: add response parameters Body.Data.EndpointUri.
+- Update API ModifyFlow: add request parameters EndpointUri.
+- Update API ModifyFlow: add response parameters Body.Data.EndpointUri.
+
+
 2026-01-08 Version: 4.0.0
 - Support API AddAddressRecoverSuspend.
 - Support API AddAuditViberOpen.